refactor: move yapayzeka implementation into chatbot plugin
