+++ docs/english-commands.md
@@ -1,0 +1,163 @@
+# İngilizce Komutlar → Türkçe Karşılıklar
+
+Aşağıdaki listede önce **senin kesin verdiğin çeviriler** işlendi. Geri kalan komutlar için de ayrı bir sütunda **öneri** sundum.
+
+> Not: Mesajındaki `blokc` ifadesini mevcut komuta göre `block` olarak uyguladım. Ayrıca `liste = liste` satırını `list -> liste` olarak yorumladım.
+
+| Komut | Türkçe karşılık | Durum |
+|---|---|---|
+| `add` | `ekle` | Kullanıcı onaylı |
+| `afk` | `uzakta` | Öneri |
+| `age` | `yaşhesap` | Kullanıcı onaylı |
+| `ai` | `yz` | Öneri |
+| `alive` | `aktif` | Öneri |
+| `allvar` | `tümayarlar` | Öneri |
+| `antibot` | `antibot` | Öneri |
+| `antidelete` | `antisilme` | Kullanıcı onaylı |
+| `antidemote` | `antiyetkidüşürme` | Kullanıcı onaylı |
+| `antifake` | `antinumara` | Kullanıcı onaylı |
+| `antilink` | `antibağlantı` | Kullanıcı onaylı |
+| `antipromote` | `antiyetkiverme` | Kullanıcı onaylı |
+| `antispam` | `antispam` | Öneri |
+| `antiword` | `antikelime` | Kullanıcı onaylı |
+| `at` | `at` | Öneri |
+| `attp` | `yazısticker` | Öneri |
+| `autodl` | `otodl` | Öneri |
+| `automute` | `otosohbetkapat` | Kullanıcı onaylı |
+| `autounmute` | `otosohbetaç` | Kullanıcı onaylı |
+| `avmix` | `sesvideobirleştir` | Öneri |
+| `bass` | `basartır` | Öneri |
+| `black` | `siyahvideo` | Öneri |
+| `blackpink` | `siyahpembe` | Öneri |
+| `block` | `engelle` | Kullanıcı onaylı |
+| `brat` | `brat` | Öneri |
+| `callreject` | `aramaengel` | Kullanıcı onaylı |
+| `chatbot` | `sohbetbotu` | Öneri |
+| `circle` | `daire` | Öneri |
+| `clear` | `sohbetsil` | Kullanıcı onaylı |
+| `cntd` | `gerisayım` | Öneri |
+| `common` | `ortaklar` | Öneri |
+| `compress` | `sıkıştır` | Öneri |
+| `dc` | `dc` | Öneri |
+| `delete` | `msjsil` | Kullanıcı onaylı |
+| `delfilter` | `filtresil` | Öneri |
+| `delsudo` | `sudosil` | Öneri |
+| `delvar` | `ayarsil` | Öneri |
+| `demote` | `yetkial` | Kullanıcı onaylı |
+| `diff` | `fark` | Öneri |
+| `doc` | `belge` | Öneri |
+| `edit` | `düzenle` | Kullanıcı onaylı |
+| `emojimix` | `emojikarışım` | Öneri |
+| `fancy` | `şekilyazı` | Öneri |
+| `fb` | `facebook` | Öneri |
+| `filter` | `filtre` | Kullanıcı onaylı |
+| `filterhelp` | `filtreyardım` | Kullanıcı onaylı |
+| `filters` | `filtreler` | Kullanıcı onaylı |
+| `flip` | `çevir` | Öneri |
+| `forward` | `msjyönlendir` | Kullanıcı onaylı |
+| `games` | `oyunlar` | Kullanıcı onaylı |
+| `gdesc` | `grupaçıklama` | Öneri |
+| `getjids` | `jidler` | Öneri |
+| `getmute` | `saatliayar` | Öneri |
+| `getstick` | `stickgetir` | Öneri |
+| `getsudo` | `sudolar` | Öneri |
+| `getvar` | `ayargetir` | Öneri |
+| `gif` | `gif` | Öneri |
+| `glock` | `grupkilit` | Öneri |
+| `gname` | `grupadı` | Öneri |
+| `goodbye` | `görüşürüz` | Öneri |
+| `gpp` | `grupfoto` | Öneri |
+| `gunlock` | `grupkilitaç` | Öneri |
+| `ig` | `igstalk` | Öneri |
+| `img` | `görsel` | Öneri |
+| `inactive` | `pasifler` | Öneri |
+| `info` | `komut` | Kullanıcı onaylı |
+| `insta` | `instagramindir` | Öneri |
+| `install` | `modülyükle` | Kullanıcı onaylı |
+| `interp` | `fpsartır` | Öneri |
+| `jid` | `jid` | Öneri |
+| `join` | `katıl` | Kullanıcı onaylı |
+| `kick` | `çıkar` | Öneri |
+| `language` | `dil` | Kullanıcı onaylı |
+| `leave` | `ayrıl` | Kullanıcı onaylı |
+| `link` | `davet` | Kullanıcı onaylı |
+| `list` | `liste` | Kullanıcı onaylı |
+| `marvel` | `marvel` | Öneri |
+| `mention` | `bahsetme` | Kullanıcı onaylı |
+| `menu` | `menü` | Kullanıcı onaylı |
+| `mode` | `mod` | Kullanıcı onaylı |
+| `mp3` | `seseçevir` | Öneri |
+| `mp4` | `videoyaçevir` | Öneri |
+| `mute` | `sohbetkapat` | Kullanıcı onaylı |
+| `naruto` | `naruto` | Öneri |
+| `pdf` | `pdf` | Öneri |
+| `pdm` | `pdm` | Öneri |
+| `photo` | `foto` | Öneri |
+| `ping` | `ping` | Öneri |
+| `pinterest` | `pinterest` | Öneri |
+| `platform` | `platform` | Öneri |
+| `plugin` | `modül` | Kullanıcı onaylı |
+| `pp` | `pp` | Öneri |
+| `promote` | `yetkiver` | Kullanıcı onaylı |
+| `pupdate` | `pupdate` | Öneri |
+| `quoted` | `alıntı` | Öneri |
+| `react` | `tepki` | Öneri |
+| `reboot` | `yenidenbaşlat` | Öneri |
+| `reload` | `yenileyükle` | Öneri |
+| `remove` | `kaldır` | Öneri |
+| `requests` | `istekler` | Kullanıcı onaylı |
+| `resize` | `boyutlandır` | Öneri |
+| `restart` | `ybaşlat` | Kullanıcı onaylı |
+| `retry` | `tekrar` | Kullanıcı onaylı |
+| `revoke` | `bağlantısıfırla` | Öneri |
+| `rotate` | `döndür` | Öneri |
+| `send` | `gönder` | Öneri |
+| `setalive` | `aktifmesaj` | Öneri |
+| `setenv` | `ortamayarla` | Öneri |
+| `setimage` | `resimayarla` | Öneri |
+| `setinfo` | `bilgiayarla` | Öneri |
+| `setname` | `adayarla` | Öneri |
+| `setowner` | `sahipayarla` | Öneri |
+| `setsudo` | `sudoayarla` | Öneri |
+| `settings` | `ayarlar` | Kullanıcı onaylı |
+| `setvar` | `ayarver` | Öneri |
+| `slow` | `yavaşlat` | Öneri |
+| `slowmo` | `ağırçekim` | Öneri |
+| `song` | `şarkı` | Öneri |
+| `soundcloud` | `soundcloud` | Öneri |
+| `sped` | `hızlandır` | Öneri |
+| `spotify` | `spotify` | Öneri |
+| `square` | `kare` | Öneri |
+| `stickcmd` | `stickerkomut` | Öneri |
+| `sticker` | `sticker` | Öneri |
+| `story` | `hikaye` | Kullanıcı onaylı |
+| `tag` | `etiketle` | Öneri |
+| `take` | `üstbilgideğiştir` | Öneri |
+| `testalive` | `aktiftest` | Öneri |
+| `testfilter` | `filtretest` | Öneri |
+| `testgay` | `gaytest` | Öneri |
+| `testgoodbye` | `görüşürüztest` | Öneri |
+| `testwelcome` | `hoşgeldintest` | Öneri |
+| `threads` | `threads` | Öneri |
+| `tiktok` | `tiktok` | Öneri |
+| `toggle` | `açkapat` | Öneri |
+| `togglefilter` | `filtreaçkapat` | Öneri |
+| `trim` | `kes` | Öneri |
+| `tts` | `metindenses` | Öneri |
+| `twitter` | `twitter` | Öneri |
+| `unblock` | `engelkaldır` | Öneri |
+| `unmute` | `sohbetaç` | Öneri |
+| `unstick` | `stickersil` | Öneri |
+| `update` | `güncelle` | Öneri |
+| `upload` | `yükle` | Öneri |
+| `uptime` | `çalışmasüresi` | Öneri |
+| `url` | `bağlantı` | Öneri |
+| `users` | `kullanıcılar` | Öneri |
+| `video` | `video` | Öneri |
+| `vmix` | `videobirleştir` | Öneri |
+| `vv` | `tekgöster` | Öneri |
+| `welcome` | `hoşgeldin` | Öneri |
+| `wiki` | `vikipedi` | Kullanıcı onaylı |
+| `yta` | `ytaudio` | Öneri |
+| `yts` | `ytara` | Öneri |
+| `ytv` | `ytvideo` | Öneri |
