+++ plugins/social.js
@@ -208,10 +208,10 @@ Module(
 
 Module(
   {
-    pattern: "story ?(.*)",
+    pattern: "hikaye ?(.*)",
     fromMe: isFromMe,
     desc: "Instagram hikaye (story) indirici",
-    usage: ".story kullanıcı adı veya bağlantı",
+    usage: ".hikaye kullanıcı adı veya bağlantı",
     use: "download",
   },
   async (message, match) => {
